# docs/swagger.yaml
basePath: /
definitions:
  github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
      request_id:
        type: string
    type: object
...
  github_com_qwerty7415963_go_be_arbitrage_internal_api.Meta:
    properties:
      cursor:
        type: string
      has_more:
        type: boolean
      limit:
        type: integer
      offset:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  github_com_qwerty7415963_go_be_arbitrage_internal_api.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
      meta:
        $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Meta'
      success:
        type: boolean
    type: object
...
  internal_execution.CreateExecutionRequest:
    properties:
      execution_mode:
        $ref: '#/definitions/internal_execution.ExecutionMode'
      intent:
        $ref: '#/definitions/internal_execution.ExecutionIntent'
      strategy_instance_id:
        type: string
    required:
    - execution_mode
    - intent
    - strategy_instance_id
    type: object
  internal_execution.Execution:
    properties:
      completed_at:
        type: string
      created_at:
        type: string
      execution_mode:
        $ref: '#/definitions/internal_execution.ExecutionMode'
      execution_plan_json: {}
      final_pnl:
        type: number
      id:
        type: string
      intent_json: {}
      opportunity_id:
        type: string
      risk_decision_id:
        type: string
      status:
        $ref: '#/definitions/internal_execution.ExecutionStatus'
      strategy_instance_id:
        type: string
      tenant_id:
        type: string
    type: object
  internal_execution.ExecutionIntent:
    properties:
      legs:
        items:
          $ref: '#/definitions/internal_execution.IntentLeg'
        type: array
      opportunity_id:
        type: string
      strategy_id:
        type: string
    type: object
  internal_execution.ExecutionLeg:
    properties:
      completed_at:
        type: string
      execution_id:
        type: string
      id:
        type: string
      instrument_id:
        type: string
      leg_index:
        type: integer
      leg_role:
        type: string
      metadata: {}
      started_at:
        type: string
      status:
        $ref: '#/definitions/internal_execution.LegStatus'
      target_notional:
        type: number
      target_quantity:
        type: number
      target_side:
        $ref: '#/definitions/internal_execution.Side'
      venue_account_id:
        type: string
    type: object
  internal_execution.ExecutionMode:
    enum:
    - PAPER
    - LIVE_MANUAL
    - LIVE_AUTO
    type: string
    x-enum-varnames:
    - ExecutionModePaper
    - ExecutionModeLiveManual
    - ExecutionModeLiveAuto
  internal_execution.ExecutionStatus:
    enum:
    - CREATED
    - RISK_APPROVED
    - AWAITING_APPROVAL
    - PLANNED
    - SUBMITTING
    - PARTIAL
    - COMPLETED
    - RECOVERY
    - FAILED
    - CANCELED
    type: string
    x-enum-varnames:
    - ExecutionStatusCreated
    - ExecutionStatusRiskApproved
    - ExecutionStatusAwaitingApproval
    - ExecutionStatusPlanned
    - ExecutionStatusSubmitting
    - ExecutionStatusPartial
    - ExecutionStatusCompleted
    - ExecutionStatusRecovery
    - ExecutionStatusFailed
    - ExecutionStatusCanceled
  internal_execution.Fill:
    properties:
      exchange_fill_id:
        type: string
      fee_amount:
        type: number
      fee_asset:
        type: string
      filled_at:
        type: string
      id:
        type: string
      liquidity_role:
        type: string
      order_id:
        type: string
      price:
        type: number
      quantity:
        type: number
      venue_account_id:
        type: string
    type: object
  internal_execution.IntentLeg:
    properties:
      instrument_id:
        type: string
      price:
        type: number
      quantity:
        type: number
      side:
        $ref: '#/definitions/internal_execution.Side'
      venue_account_id:
        type: string
    type: object
  internal_execution.LegStatus:
    enum:
    - PENDING
    - SUBMITTING
    - OPEN
    - PARTIAL
    - FILLED
    - FAILED
    - CANCELED
    - RECOVERY
    type: string
    x-enum-varnames:
    - LegStatusPending
    - LegStatusSubmitting
    - LegStatusOpen
    - LegStatusPartial
    - LegStatusFilled
    - LegStatusFailed
    - LegStatusCanceled
    - LegStatusRecovery
  internal_execution.Side:
    enum:
    - BUY
    - SELL
    type: string
    x-enum-varnames:
    - SideBuy
    - SideSell
...
host: localhost:8080
info:
  contact:
    email: support@arbitrage.com
    name: API Support
  description: Backend API for arbitrage trading platform
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Arbitrage Platform API
  version: "1.0"
paths:
...
  /api/v1/executions:
    get:
      description: Executions for the caller's tenant (currently always an empty array
        — storage not wired). Requires JWT + admin role (401 AUTH-003 without token,
        403 AUTH-005 for non-admin).
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_execution.Execution'
                  type: array
              type: object
        "401":
          description: AUTH-003
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "403":
          description: AUTH-005
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
      security:
      - BearerAuth: []
      summary: List executions
      tags:
      - executions
    post:
      consumes:
      - application/json
      description: Creates an execution scoped to the caller's tenant. Requires JWT
        + admin role.
      parameters:
      - description: Execution definition
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_execution.CreateExecutionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_execution.Execution'
              type: object
        "400":
          description: COMMON-902
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "401":
          description: AUTH-003
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "403":
          description: AUTH-005
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "500":
          description: COMMON-901
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
      security:
      - BearerAuth: []
      summary: Create execution
      tags:
      - executions
  /api/v1/executions/{id}:
    get:
      description: One execution by ID. Requires JWT + admin role.
      parameters:
      - description: Execution ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_execution.Execution'
              type: object
        "400":
          description: COMMON-902
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "401":
          description: AUTH-003
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "403":
          description: AUTH-005
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "404":
          description: COMMON-903
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
      security:
      - BearerAuth: []
      summary: Get execution
      tags:
      - executions
  /api/v1/executions/{id}/cancel:
    post:
      description: Cancels an execution. Requires JWT + admin role.
      parameters:
      - description: Execution ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: COMMON-902
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "401":
          description: AUTH-003
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "403":
          description: AUTH-005
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "500":
          description: COMMON-901
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
      security:
      - BearerAuth: []
      summary: Cancel execution
      tags:
      - executions
  /api/v1/executions/{id}/fills:
    get:
      description: Fills for one execution; empty array when none. Requires JWT +
        admin role.
      parameters:
      - description: Execution ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_execution.Fill'
                  type: array
              type: object
        "400":
          description: COMMON-902
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "401":
          description: AUTH-003
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "403":
          description: AUTH-005
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "500":
          description: COMMON-901
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
      security:
      - BearerAuth: []
      summary: List execution fills
      tags:
      - executions
  /api/v1/executions/{id}/legs:
    get:
      description: Execution legs (per-venue splits) for one execution; empty array
        when none. Requires JWT + admin role.
      parameters:
      - description: Execution ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_execution.ExecutionLeg'
                  type: array
              type: object
        "400":
          description: COMMON-902
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "401":
          description: AUTH-003
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "403":
          description: AUTH-005
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "500":
          description: COMMON-901
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
      security:
      - BearerAuth: []
      summary: List execution legs
      tags:
      - executions
  /api/v1/executions/{id}/submit:
    post:
      description: Submits an execution to its venue adapter. Requires JWT + admin
        role.
      parameters:
      - description: Execution ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: COMMON-902
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "401":
          description: AUTH-003
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "403":
          description: AUTH-005
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "500":
          description: COMMON-901
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
      security:
      - BearerAuth: []
      summary: Submit execution
      tags:
      - executions
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"