# internal/execution/handler.go
package execution

import (
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/google/uuid"
	"github.com/qwerty7415963/go_be_arbitrage/internal/api"
	"github.com/qwerty7415963/go_be_arbitrage/internal/domain"
)

type Handler struct {
	service *Service
}

func NewHandler(service *Service) *Handler {
	return &Handler{service: service}
}

func (h *Handler) RegisterRoutes(router *gin.RouterGroup, jwtMiddleware ...gin.HandlerFunc) {
	executions := router.Group("/executions")
	if len(jwtMiddleware) > 0 {
		executions.Use(jwtMiddleware[0])
	}
	executions.Use(func(c *gin.Context) {
		role, exists := c.Get("role")
		if !exists {
			api.RespondError(c, domain.NewError(domain.ErrCodeAuthForbidden, "role not found in token"))
			c.Abort()
			return
		}
		roleStr, ok := role.(string)
		if !ok {
			api.RespondError(c, domain.NewError(domain.ErrCodeAuthForbidden, "invalid role type"))
			c.Abort()
			return
		}
		if roleStr != "admin" {
			api.RespondError(c, domain.NewError(domain.ErrCodeAuthForbidden, "insufficient permissions"))
			c.Abort()
			return
		}
		c.Next()
	})
	{
		executions.GET("", h.List)
		executions.POST("", h.Create)
		executions.GET("/:id", h.GetByID)
		executions.POST("/:id/submit", h.Submit)
		executions.POST("/:id/cancel", h.Cancel)
		executions.GET("/:id/legs", h.ListLegs)
		executions.GET("/:id/fills", h.ListFills)
	}
}

func (h *Handler) getTenantID(c *gin.Context) (uuid.UUID, bool) {
	tenantIDStr, exists := c.Get("tenant_id")
	if !exists {
		api.RespondError(c, domain.NewError(domain.ErrCodeAuthForbidden, "tenant not found"))
		return uuid.Nil, false
	}
	s, ok := tenantIDStr.(string)
	if !ok {
		api.RespondError(c, domain.NewError(domain.ErrCodeAuthForbidden, "invalid tenant"))
		return uuid.Nil, false
	}
	tenantID, err := uuid.Parse(s)
	if err != nil {
		api.RespondError(c, domain.NewError(domain.ErrCodeAuthForbidden, "invalid tenant"))
		return uuid.Nil, false
	}
	return tenantID, true
}

// List godoc
// @Summary      List executions
// @Description  Executions for the caller's tenant (currently always an empty array — storage not wired). Requires JWT + admin role (401 AUTH-003 without token, 403 AUTH-005 for non-admin).
// @Tags         executions
// @Produce      json
// @Security     BearerAuth
// @Success      200  {object}  api.Response{data=[]Execution}
// @Failure      401  {object}  api.Response{error=api.ErrorBody}  "AUTH-003"
// @Failure      403  {object}  api.Response{error=api.ErrorBody}  "AUTH-005"
// @Router       /api/v1/executions [get]
func (h *Handler) List(c *gin.Context) {
	api.RespondList(c, []*Execution{}, nil)
}

// Create godoc
// @Summary      Create execution
// @Description  Creates an execution scoped to the caller's tenant. Requires JWT + admin role.
// @Tags         executions
// @Accept       json
// @Produce      json
// @Security     BearerAuth
// @Param        request  body      CreateExecutionRequest  true  "Execution definition"
// @Success      201  {object}  api.Response{data=Execution}
// @Failure      400  {object}  api.Response{error=api.ErrorBody}  "COMMON-902"
// @Failure      401  {object}  api.Response{error=api.ErrorBody}  "AUTH-003"
// @Failure      403  {object}  api.Response{error=api.ErrorBody}  "AUTH-005"
// @Failure      500  {object}  api.Response{error=api.ErrorBody}  "COMMON-901"
// @Router       /api/v1/executions [post]
func (h *Handler) Create(c *gin.Context) {
	tenantID, ok := h.getTenantID(c)
	if !ok {
		return
	}
	var req CreateExecutionRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		api.RespondError(c, domain.NewError(domain.ErrCodeValidation, "invalid request body"))
		return
	}
	exec, err := h.service.CreateExecution(c.Request.Context(), tenantID, &req)
	if err != nil {
		api.RespondError(c, domain.NewError(domain.ErrCodeInternal, err.Error()))
		return
	}
	c.JSON(http.StatusCreated, api.Response{Success: true, Data: exec})
}

// GetByID godoc
// @Summary      Get execution
// @Description  One execution by ID. Requires JWT + admin role.
// @Tags         executions
// @Produce      json
// @Security     BearerAuth
// @Param        id   path      string  true  "Execution ID"
// @Success      200  {object}  api.Response{data=Execution}
// @Failure      400  {object}  api.Response{error=api.ErrorBody}  "COMMON-902"
// @Failure      401  {object}  api.Response{error=api.ErrorBody}  "AUTH-003"
// @Failure      403  {object}  api.Response{error=api.ErrorBody}  "AUTH-005"
// @Failure      404  {object}  api.Response{error=api.ErrorBody}  "COMMON-903"
// @Router       /api/v1/executions/{id} [get]
func (h *Handler) GetByID(c *gin.Context) {
	id, err := uuid.Parse(c.Param("id"))
	if err != nil {
		api.RespondError(c, domain.NewError(domain.ErrCodeValidation, "invalid execution ID"))
		return
	}
	exec, err := h.service.GetExecution(c.Request.Context(), id)
	if err != nil {
		api.RespondError(c, domain.NewError(domain.ErrCodeNotFound, "execution not found"))
		return
	}
	api.RespondSuccess(c, exec)
}

// Submit godoc
// @Summary      Submit execution
// @Description  Submits an execution to its venue adapter. Requires JWT + admin role.
// @Tags         executions
// @Produce      json
// @Security     BearerAuth
// @Param        id   path      string  true  "Execution ID"
// @Success      200  {object}  api.Response{data=map[string]string}
// @Failure      400  {object}  api.Response{error=api.ErrorBody}  "COMMON-902"
// @Failure      401  {object}  api.Response{error=api.ErrorBody}  "AUTH-003"
// @Failure      403  {object}  api.Response{error=api.ErrorBody}  "AUTH-005"
// @Failure      500  {object}  api.Response{error=api.ErrorBody}  "COMMON-901"
// @Router       /api/v1/executions/{id}/submit [post]
func (h *Handler) Submit(c *gin.Context) {
	id, err := uuid.Parse(c.Param("id"))
	if err != nil {
		api.RespondError(c, domain.NewError(domain.ErrCodeValidation, "invalid execution ID"))
		return
	}
	if err := h.service.SubmitExecution(c.Request.Context(), id); err != nil {
		api.RespondError(c, domain.NewError(domain.ErrCodeInternal, err.Error()))
		return
	}
	api.RespondSuccess(c, gin.H{"message": "execution submitted"})
}

// Cancel godoc
// @Summary      Cancel execution
// @Description  Cancels an execution. Requires JWT + admin role.
// @Tags         executions
// @Produce      json
// @Security     BearerAuth
// @Param        id   path      string  true  "Execution ID"
// @Success      200  {object}  api.Response{data=map[string]string}
// @Failure      400  {object}  api.Response{error=api.ErrorBody}  "COMMON-902"
// @Failure      401  {object}  api.Response{error=api.ErrorBody}  "AUTH-003"
// @Failure      403  {object}  api.Response{error=api.ErrorBody}  "AUTH-005"
// @Failure      500  {object}  api.Response{error=api.ErrorBody}  "COMMON-901"
// @Router       /api/v1/executions/{id}/cancel [post]
func (h *Handler) Cancel(c *gin.Context) {
	id, err := uuid.Parse(c.Param("id"))
	if err != nil {
		api.RespondError(c, domain.NewError(domain.ErrCodeValidation, "invalid execution ID"))
		return
	}
	if err := h.service.CancelExecution(c.Request.Context(), id); err != nil {
		api.RespondError(c, domain.NewError(domain.ErrCodeInternal, err.Error()))
		return
	}
	api.RespondSuccess(c, gin.H{"message": "execution canceled"})
}

// ListLegs godoc
// @Summary      List execution legs
// @Description  Execution legs (per-venue splits) for one execution; empty array when none. Requires JWT + admin role.
// @Tags         executions
// @Produce      json
// @Security     BearerAuth
// @Param        id   path      string  true  "Execution ID"
// @Success      200  {object}  api.Response{data=[]ExecutionLeg}
// @Failure      400  {object}  api.Response{error=api.ErrorBody}  "COMMON-902"
// @Failure      401  {object}  api.Response{error=api.ErrorBody}  "AUTH-003"
// @Failure      403  {object}  api.Response{error=api.ErrorBody}  "AUTH-005"
// @Failure      500  {object}  api.Response{error=api.ErrorBody}  "COMMON-901"
// @Router       /api/v1/executions/{id}/legs [get]
func (h *Handler) ListLegs(c *gin.Context) {
	id, err := uuid.Parse(c.Param("id"))
	if err != nil {
		api.RespondError(c, domain.NewError(domain.ErrCodeValidation, "invalid execution ID"))
		return
	}
	legs, err := h.service.ListLegs(c.Request.Context(), id)
	if err != nil {
		api.RespondError(c, domain.NewError(domain.ErrCodeInternal, err.Error()))
		return
	}
	if legs == nil {
		legs = []*ExecutionLeg{}
	}
	api.RespondList(c, legs, nil)
}

// ListFills godoc
// @Summary      List execution fills
// @Description  Fills for one execution; empty array when none. Requires JWT + admin role.
// @Tags         executions
// @Produce      json
// @Security     BearerAuth
// @Param        id   path      string  true  "Execution ID"
// @Success      200  {object}  api.Response{data=[]Fill}
// @Failure      400  {object}  api.Response{error=api.ErrorBody}  "COMMON-902"
// @Failure      401  {object}  api.Response{error=api.ErrorBody}  "AUTH-003"
// @Failure      403  {object}  api.Response{error=api.ErrorBody}  "AUTH-005"
// @Failure      500  {object}  api.Response{error=api.ErrorBody}  "COMMON-901"
// @Router       /api/v1/executions/{id}/fills [get]
func (h *Handler) ListFills(c *gin.Context) {
	id, err := uuid.Parse(c.Param("id"))
	if err != nil {
		api.RespondError(c, domain.NewError(domain.ErrCodeValidation, "invalid execution ID"))
		return
	}
	fills, err := h.service.ListFills(c.Request.Context(), id)
	if err != nil {
		api.RespondError(c, domain.NewError(domain.ErrCodeInternal, err.Error()))
		return
	}
	if fills == nil {
		fills = []*Fill{}
	}
	api.RespondList(c, fills, nil)
}